Long2Short/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
altair==4.2.2
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.5
bertopic==0.17.3
bitsandbytes==0.45.5
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
clean-fid==0.1.35
click==8.1.8
click-default-group==1.2.4
clip-anytorch==2.6.0
colorama==0.4.6
colorspacious==1.1.2
contourpy==1.3.2
controlnet_aux==0.0.7
cycler==0.12.1
daltonize==0.2.0
dateparser==1.2.2
dctorch==0.1.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
diffusers @ git+https://github.com/huggingface/diffusers.git@b4be42282dc9bf9deccc2f60f5db952de772cf42
docker-pycreds==0.4.0
easygui==0.98.3
einops==0.7.0
entrypoints==0.4
fake-useragent==2.2.0
fastapi==0.116.2
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
flatten-json==0.1.14
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.46.1
gradio_client==1.13.1
gradio_logsview @ https://huggingface.co/spaces/cocktailpeanut/gradio_logsview/resolve/main/gradio_logsview-0.0.17-py3-none-any.whl#sha256=c595d1ab65cc04b009b538652f64dcdb2e9d30491ee0ffc12f25101da5e77042
greenlet==3.2.4
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h2==4.3.0
hdbscan==0.8.40
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
invisible-watermark==0.2.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.5.2
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
k-diffusion==0.1.1.post1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lap==0.5.12
lazy_loader==0.4
-e git+https://github.com/kohya-ss/sd-scripts@b11c053b8fcd1c4532dc3a37e70109e08aafa2ec#egg=library
lightning-utilities==0.14.3
lion-pytorch==0.0.6
llvmlite==0.44.0
lpips==0.1.4
lxml==6.0.1
lycoris_lora==1.8.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
mss==10.1.0
multidict==6.4.3
narwhals==2.5.0
networkx==3.4.2
ngrok==1.5.1
ninja==1.11.1.4
nltk==3.9.1
nodriver==0.47.0
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.32.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optimum-quanto==0.2.7
orjson==3.10.16
outcome==1.3.0.post0
oyaml==1.0
packaging==25.0
pandas==2.1.4
pefile==2023.2.7
peft==0.15.2
pillow==11.3.0
platformdirs==4.3.7
playwright==1.55.0
plotly==6.3.0
pluggy==1.6.0
polars==1.33.1
praw==7.7.1
prawcore==2.4.0
prodigy-plus-schedule-free==1.9.0
prodigyopt==1.1.2
proglog==0.1.12
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-fid==0.3.0
pytorch-lightning==1.9.0
pytorch_optimizer==3.5.0
pytube==15.0.0
pytz==2025.2
PyWavelets==1.8.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.7.0
rpds-py==0.24.0
ruff==0.11.6
safehttpx==0.1.6
safetensors==0.5.3
schedulefree==1.4
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.8
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
stringzilla==3.12.5
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tifffile==2025.3.30
TikTokApi==7.1.0
tiktokapipy==0.2.4.post2
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchdiffeq==0.2.5
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.22.0+cu118
tqdm==4.67.1
trampoline==0.1.2
transformers==4.48.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.200
ultralytics-thop==2.0.17
umap-learn==0.5.9.post2
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.34.2
voluptuous==0.13.1
wandb==0.19.10
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.0
youtube-comment-downloader==0.1.78
youtube-transcript-api==1.2.2
yt-dlp==2025.9.5
zipp==3.21.0
